refactors and  renames function-services folder
