+++ packages/models-library/src/models_library/function_services_catalog/constants.py
