+++ packages/models-library/src/models_library/function_services_catalog/services/nodes_group.py
